requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
av==13.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-default-group==1.2.4
cloup==3.0.7
colour==0.1.5
contourpy==1.3.0
cssselect==1.3.0
cycler==0.12.1
Cython==3.1.0
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.0.2
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
glcontext==3.0.0
google-api-core==2.25.0rc1
google-api-python-client==2.118.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
islenska==1.0.3
isosurfaces==0.1.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
kiwisolver==1.4.7
lxml==4.9.3
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
mpmath==1.3.0
multidict==6.4.3
multipledispatch==1.0.0
networkx==2.8.8
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.78.0
packaging==25.0
pandas==2.1.4
pathlib==1.0.1
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.26.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyparsing==3.2.3
pyrr==0.10.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
reynir==3.5.7
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
schedule==1.2.1
scikit-learn==1.6.1
scipy==1.13.1
screeninfo==0.8.1
seaborn==0.13.2
sentence-transformers==4.1.0
sgmllib3k==1.0.0
six==1.17.0
skia-pathops==0.7.4
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
srt==3.5.3
svgelements==1.9.6
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizer==3.4.5
tokenizers==0.21.1
torch==2.1.2
tqdm==4.66.2
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
watchdog==2.3.1
Werkzeug==3.1.3
yarl==1.20.0
zipp==3.21.0
